+++ audio_synch_tool/__main__.py
@@ -11,7 +11,7 @@ import soundfile as sf
 import torch
 
 # from .utils import Timestamp
-from .plotters import MultipleDownsampledPlotter1D
+from .plotters import MultipleDownsampledPlotter1D, AudioMvnSynchTool
 from .plotters import TextPrompt, ShiftRightTool, StretchRightTool
 from .mvn import Mvn
 
@@ -52,12 +52,13 @@ NUM_XTICKS = 20
 
 # define paths
 WAV_PATH = os.path.join(os.getenv("HOME"), "SAG_D1_10_M-Jem-2.wav")
+
 MVN_PATH = os.path.join(os.getenv("HOME"),
                         "SAG_D1-003_(snippet)_slate01_2-23-29.701.mvnx")
-# MVN_PATH = os.path.join(os.getenv("HOME"),
-#                       "SAG_D1-003_(snippet)_slate03_2-35-44.759.mvnx")
+
 MVN_SCHEMA_PATH = os.path.join(os.path.dirname(os.path.abspath(__file__)),
                                "data", "mvn_schema_adapted.xsd")
+
 
 # load fiels
 wav_arr, audio_samplerate = sf.read(WAV_PATH)
@@ -65,38 +66,30 @@ mocap = Mvn(MVN_PATH, MVN_SCHEMA_PATH)
 
 # extract info from mocap
 mocap_samplerate = float(mocap.mvn.subject.attrib["frameRate"])
-mocap_segments = mocap.get_segments()
-frame_sequences = mocap.get_normalframe_sequences("cpu")
+mocap_segments = mocap.extract_segments()
+frames_metadata, config_frames, normal_frames = mocap.extract_frame_info()
+frame_sequences = mocap.extract_normalframe_sequences(frames_metadata,
+                                                      normal_frames, "cpu")
 mocap_accelerations_3d = frame_sequences["acceleration"]
 mocap_accel_norm = torch.norm(mocap_accelerations_3d, 2, dim=-1)
 
+mocap.set_audio_synch(10000, 2000000)
+# mocap.export("quack.asdf", extra_comment="Bound to audio file XYZ")
+# input("asdfasdf")
 
 
 
-# plot
-y_arrays = [[wav_arr],
-            [mocap_accel_norm[:, mocap_segments.index("LeftShoulder")].numpy(),
-             mocap_accel_norm[:, mocap_segments.index("LeftForeArm")].numpy(),
-             mocap_accel_norm[:, mocap_segments.index("LeftHand")].numpy()],
-            [mocap_accel_norm[:, mocap_segments.index("RightShoulder")].numpy(),
-             mocap_accel_norm[:, mocap_segments.index("RightForeArm")].numpy(),
-             mocap_accel_norm[:, mocap_segments.index("RightHand")].numpy()]]
 
-x_arrays = [[torch.arange(len(yarr)).numpy() for yarr in yarrs]
-            for yarrs in y_arrays]
 
-# x_arrays[0][0] -= 8000 * 60
+# {'time': '0', 'tc': '00:00:00:000', 'ms': '0', 'type': 'identity'}
+# {'time': '0', 'tc': '00:00:00:000', 'ms': '0', 'type': 'tpose'}
+# {'time': '0', 'tc': '00:00:00:000', 'ms': '0', 'type': 'tpose-isb'}
+# {'time': '0', 'index': '0', 'tc': '02:23:28:164', 'ms': '1515983008686', 'type': 'normal'}
+# {'time': '5', 'index': '1', 'tc': '02:23:28:165', 'ms': '1515983008691', 'type': 'normal'}
+# {'time': '9', 'index': '2', 'tc': '02:23:28:166', 'ms': '1515983008695', 'type': 'normal'}
+# {'time': '13', 'index': '3', 'tc': '02:23:28:167', 'ms': '1515983008699', 'type': 'normal'}
+# {'time': '17', 'index': '4', 'tc': '02:23:28:168', 'ms': '1515983008703', 'type': 'normal'}
 
-for a in x_arrays[1]:
-    a *= 50000
-for a in x_arrays[2]:
-    a *= 50000
-
-samplerates = [audio_samplerate, mocap_samplerate, mocap_samplerate * 2]
-# samplerates = [audio_samplerate, mocap_samplerate, mocap_samplerate]
-tied_plots = [True, True, True] # [False, False, False]
-p = MultipleDownsampledPlotter1D(y_arrays, samplerates, MAX_SAMPLES_PLOTTED,
-                                 tied_plots, x_arrays)
 
 
 # mvn_arrays = [[mocap_accel_norm[:, mocap_segments.index("LeftShoulder")].numpy(),
@@ -105,13 +98,63 @@ p = MultipleDownsampledPlotter1D(y_arrays, samplerates, MAX_SAMPLES_PLOTTED,
 #               [mocap_accel_norm[:, mocap_segments.index("RightShoulder")].numpy(),
 #                mocap_accel_norm[:, mocap_segments.index("RightForeArm")].numpy(),
 #                mocap_accel_norm[:, mocap_segments.index("RightHand")].numpy()]]
+# p = AudioMvnSynchTool(wav_arr, mvn_arrays, audio_samplerate,
+#                       mocap_samplerate, MAX_SAMPLES_PLOTTED)
 
-# p = AudioMvnSynchTool(wav_arr, mvn_arrays, audio_samplerate, mocap_samplerate,
-#                       MAX_SAMPLES_PLOTTED)
+# mocap.set_audio_synch(10000, 2000000)
+p = AudioMvnSynchTool(wav_arr, audio_samplerate, mocap, MAX_SAMPLES_PLOTTED)
 
-textbox_widget = TextPrompt
-toolbar_widgets = ShiftRightTool, StretchRightTool
+# textbox_widget = TextPrompt
+# toolbar_widgets = ShiftRightTool, StretchRightTool
+# fig = p.make_fig(textbox_widget, toolbar_widgets)
 
 
-fig = p.make_fig(textbox_widget, toolbar_widgets)
+fig = p.make_fig()
 plt.show()
+
+
+# # fig = p.make_fig(textbox_widget, toolbar_widgets)
+# fig = p.make_fig()
+# plt.show()
+
+
+
+
+# # plot
+# y_arrays = [[wav_arr],
+#             [mocap_accel_norm[:, mocap_segments.index("LeftShoulder")].numpy(),
+#              mocap_accel_norm[:, mocap_segments.index("LeftForeArm")].numpy(),
+#              mocap_accel_norm[:, mocap_segments.index("LeftHand")].numpy()],
+#             [mocap_accel_norm[:, mocap_segments.index("RightShoulder")].numpy(),
+#              mocap_accel_norm[:, mocap_segments.index("RightForeArm")].numpy(),
+#              mocap_accel_norm[:, mocap_segments.index("RightHand")].numpy()]]
+
+
+
+
+
+
+# x_arrays = [[torch.arange(len(yarr)).numpy() for yarr in yarrs]
+#             for yarrs in y_arrays]
+
+# # x_arrays[0][0] -= 8000 * 60
+
+# for a in x_arrays[1]:
+#     a *= 50000
+# for a in x_arrays[2]:
+#     a *= 50000
+
+# samplerates = [audio_samplerate, mocap_samplerate, mocap_samplerate * 2]
+# # samplerates = [audio_samplerate, mocap_samplerate, mocap_samplerate]
+# tied_plots = [True, True, True] # [False, False, False]
+# p = MultipleDownsampledPlotter1D(y_arrays, samplerates, MAX_SAMPLES_PLOTTED,
+#                                  tied_plots, x_arrays)
+
+
+# textbox_widget = TextPrompt
+# toolbar_widgets = ShiftRightTool, StretchRightTool
+
+
+# # fig = p.make_fig(textbox_widget, toolbar_widgets)
+# fig = p.make_fig()
+# plt.show()