+++ CHANGELOG.md
@@ -42,13 +42,15 @@ Started dev
 
 * Allow pseudo-shared axes with different formatters DONE
 
-* Add widgets
+* Add widgets DONE
 
-* Add multi-samplerate label support: right now different "samplerates" aren't really supported: we have to multiply the x axis and this ends up showing the wrong sample number on the x axis. Ideally, all samplerate/shared combinations are possible and accurate.
+* Add multi-samplerate label support: right now different "samplerates" aren't really supported: we have to multiply the x axis and this ends up showing the wrong sample number on the x axis. Ideally, all samplerate/shared combinations are possible and accurate. right now the MVN signal is "adapted" to the audio freq. the buttons shift all the signals because they grab the "shared" listeners... this is also messy. need a plan to proceed further!
 
 * mvn processing facilities:
+  * serialization DONE
   * Fix timestamp
   * Shift and stretch
+
 
 ### low priority
 * replace all np dependencies with torch DONE